requirements.txt
mkdocs>=1.6.1
mkdocs-material>=9.6.20
mkdocs-macros-plugin>=1.3.9
Jinja2>=3.1.6
PyYAML>=6.0
python-libsbml>=5.20.5
sbmlpbkutils @ git+https://github.com/jwkruisselbrink/sbml-pbk-utils.git@v0.31.0
